Takes Investment to Transaction Generators
