+++ app/modules/investmentFlow/sagas.ts
@@ -4,8 +4,6 @@ import { put, select, takeEvery, takeLatest } from "redux-saga/effects";
 
 import { TGlobalDependencies } from "../../di/setupBindings";
 import { ETOCommitment } from "../../lib/contracts/ETOCommitment";
-import { ContractsService } from "../../lib/web3/ContractsService";
-import { ITxData } from "../../lib/web3/Web3Manager";
 import { IAppState } from "../../store";
 import { addBigNumbers, compareBigNumbers } from "../../utils/BigNumberUtils";
 import { convertToBigInt } from "../../utils/Money.utils";
@@ -21,7 +19,6 @@ import {
   selectLockedEtherBalance,
   selectLockedEuroTokenBalance,
 } from "../wallet/selectors";
-import { selectEthereumAddressWithChecksum } from "../web3/selectors";
 import {
   EBankTransferFlowState,
   EInvestmentCurrency,
@@ -157,7 +154,7 @@ function* start(action: TAction): any {
   yield setGasPrice();
 }
 
-function* onTxModalHide(): any {
+export function* onInvestmentTxModalHide(): any {
   const state: IInvestmentFlowState = yield select((s: IAppState) => s.investmentFlow);
   if (!selectIsBankTransferModalOpened(state)) {
     yield put(actions.investmentFlow.resetInvestment());
@@ -168,107 +165,6 @@ function* setGasPrice(): any {
   const gas: IGasState = yield select((s: IAppState) => s.gas);
   yield put(actions.investmentFlow.setGasPrice(gas.gasPrice && gas.gasPrice.standard));
   yield put(actions.investmentFlow.validateInputs());
-}
-
-function createTxData(
-  state: IAppState,
-  txInput: string,
-  contractAddress: string,
-): ITxData | undefined {
-  return {
-    to: contractAddress,
-    from: selectEthereumAddressWithChecksum(state.web3),
-    data: txInput,
-    value: "0",
-    gas: state.investmentFlow.gasAmount,
-    gasPrice: state.investmentFlow.gasPrice,
-  };
-}
-
-function getEtherLockTransaction(
-  state: IAppState,
-  contractsService: ContractsService,
-  etoId: string,
-): ITxData | undefined {
-  const txInput = contractsService.etherLock
-    .transferTx(etoId, new BigNumber(state.investmentFlow.ethValueUlps), [""])
-    .getData();
-  return createTxData(state, txInput, contractsService.etherLock.address);
-}
-
-function getEuroLockTransaction(
-  state: IAppState,
-  contractsService: ContractsService,
-  etoId: string,
-): ITxData | undefined {
-  const txInput = contractsService.euroLock
-    .transferTx(etoId, new BigNumber(state.investmentFlow.euroValueUlps), [""])
-    .getData();
-  return createTxData(state, txInput, contractsService.euroLock.address);
-}
-
-function getEtherTokenTransaction(
-  state: IAppState,
-  contractsService: ContractsService,
-  etoId: string,
-): ITxData | undefined {
-  const etherTokenBalance = state.wallet.data!.etherTokenBalance;
-  const i = state.investmentFlow;
-  let txDetails: ITxData | undefined;
-
-  // transaction can be fully covered by etherTokens
-  if (compareBigNumbers(etherTokenBalance, i.ethValueUlps) >= 0) {
-    // need to call 3 args version of transfer method. See the abi in the contract.
-    // so we call the rawWeb3Contract directly
-    const txInput = contractsService.etherToken.rawWeb3Contract.transfer[
-      "address,uint256,bytes"
-    ].getData(etoId, i.ethValueUlps, "");
-    txDetails = createTxData(state, txInput, contractsService.etherToken.address);
-
-    // fill up etherToken with ether from walle}t
-  } else {
-    const ethVal = new BigNumber(i.ethValueUlps);
-    const difference = ethVal.sub(etherTokenBalance);
-    const txCall = contractsService.etherToken.depositAndTransferTx(etoId, ethVal, [""]);
-    txDetails = {
-      to: contractsService.etherToken.address,
-      from: selectEthereumAddressWithChecksum(state.web3),
-      data: txCall.getData(),
-      value: difference.toString(),
-      gas: i.gasAmount,
-      gasPrice: i.gasPrice,
-    };
-  }
-
-  return txDetails;
-}
-
-function* generateTransaction({ contractsService }: TGlobalDependencies): any {
-  const state: IAppState = yield select();
-  const i = state.investmentFlow;
-  const eto = selectEtoById(state.publicEtos, i.etoId);
-
-  if (!eto || !selectReadyToInvest(i)) {
-    throw new Error("Investment data is not valid to create an Transaction");
-  }
-
-  let txDetails: ITxData | undefined;
-
-  switch (i.investmentType) {
-    case EInvestmentType.InvestmentWallet:
-      txDetails = getEtherTokenTransaction(state, contractsService, eto.etoId);
-      break;
-    case EInvestmentType.ICBMEth:
-      txDetails = getEtherLockTransaction(state, contractsService, eto.etoId);
-      break;
-    case EInvestmentType.ICBMnEuro:
-      txDetails = getEuroLockTransaction(state, contractsService, eto.etoId);
-      break;
-  }
-
-  if (txDetails) {
-    yield put(actions.txSender.txSenderAcceptDraft(txDetails));
-  }
 }
 
 function* recalculateCurrencies(): any {
@@ -301,10 +197,6 @@ export function* investmentFlowSagas(): any {
   yield takeEvery("INVESTMENT_FLOW_START", start);
   yield takeEvery("INVESTMENT_FLOW_SHOW_BANK_TRANSFER_SUMMARY", showBankTransferSummary);
   yield takeEvery("INVESTMENT_FLOW_SHOW_BANK_TRANSFER_DETAILS", showBankTransferDetails);
-  yield takeEvery("TX_SENDER_HIDE_MODAL", onTxModalHide);
   yield takeEvery("GAS_API_LOADED", setGasPrice);
   yield takeEvery("TOKEN_PRICE_SAVE", recalculateCurrencies);
-
-  // TODO: Take out of investment
-  yield takeEvery("TX_SENDER_HIDE_MODAL", stop);
 }
